requirements.txt
pandas==1.0.3
numpy==1.21.4
Pillow==8.4.0
opencv-python==4.5.4.60